Examination Timetabling
we assume that only one room with finite seating capacity is available
for timetabling and five timeslots are available every day. Here is the
list of hard and soft constraints:

Hard Constraints:

Each course has one exam, which is conducted only once, and all students
enrolled for the course must take the exam in the same time slot.
The first day in the timetable is 1 and there are five time slots {1, 2,
3, 4, 5} available every day.
There are no holidays! Also, there must be no unused timeslots between
the first and the last day of exams in the timetable.
Exams with one or more common students cannot be assigned to the same
time slot, i.e. a student cannot take more than one exam in the same
time slot.
Multiple exams can be assigned to a single time slot.
Total number of students taking exams in a given time slot should not
exceed the room capacity specified in the problem.
The total number of time slots used by the timetable should not exceed
the maximum number of timeslots specified in the problem.
Soft Constraints (Objective Functions):

Our version of the problem will allow for one of the following two
objective functions:
Total number of timeslots used by the timetable should be minimized.
Total student cost should be minimized